redesign the header and make it responsive
