+++ packages/webview-common/src/components/icon.tsx
@@ -1,0 +1,14 @@
+export const InfoIcon = () => (
+  <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor">
+    <path
+      fill="#fff"
+      d="M8 16a8 8 0 1 1 8-8 8.009 8.009 0 0 1-8 8ZM8 1a7 7 0 1 0 7 7 7.008 7.008 0 0 0-7-7Zm-.434 4.564a.846.846 0 0 1-.312-.314.888.888 0 0 1 0-.872.831.831 0 0 1 .312-.31.88.88 0 0 1 1.187.313c.078.13.119.28.117.432A.865.865 0 0 1 8 5.68a.85.85 0 0 1-.434-.116Zm1.083 6.77H7.36V6.5h1.291l-.002 5.835Z"
+    />
+  </svg>
+);
+
+export const HamburgerIcon = () => (
+  <svg xmlns="http://www.w3.org/2000/svg" width="14" height="11" fill="currentColor">
+    <path d="M13.667 6.333H.333v-1h13.334v1Zm0-5.666H.333v1h13.334v-1Zm0 9.333H.333v1h13.334v-1Z" />
+  </svg>
+);
